% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-dnam.R
\name{beta2M}
\alias{beta2M}
\title{Convert beta values to M values}
\usage{
beta2M(x, alpha = 1e-06)
}
\arguments{
\item{x}{scalar, matrix, or data frame of beta values}

\item{alpha}{a small number to avoid log(0), default is 1e-6}
}
\value{
M values
}
\description{
Convert beta values to M values
}
\examples{
beta2M(meth_df)
}
