% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{check_packages}
\alias{check_packages}
\title{Check which packages are not installed}
\usage{
check_packages(pkgs)
}
\arguments{
\item{pkgs}{Character vector of package names.}
}
\value{
Character vector of packages not currently available.
}
\description{
Check which packages are not installed
}
\examples{
check_packages(c("stats","definitely_not_a_real_pkg"))
}
