% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{closestMatch}
\alias{closestMatch}
\title{find the closest match of a string in a vector of strings}
\usage{
closestMatch(string, stringVector)
}
\arguments{
\item{string}{A string}

\item{stringVector}{A vector of strings to be compared with}
}
\value{
a list of the closest match and its index in the vector
}
\description{
This function searches the closest match of a string in a vector of strings using the levenshtein distance, sometimes useful for matching gene names.
}
\examples{
closestMatch("DFNB31", c("DNMT1", "DTNBP1", "IFNB1"))
#[[1]] [1] "DTNBP1" "IFNB1"
#[[2]] [1] 2 3
}
