% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{createNamedList}
\alias{createNamedList}
\title{Create a named list}
\usage{
createNamedList(vec)
}
\arguments{
\item{vec}{A character vector specifying names of each element in the list.}
}
\value{
A named list with the same length as \code{vec}.
}
\description{
Given a character vector, this function creates an empty list with names
corresponding to the elements of the vector.
}
\examples{
createNamedList(c("x", "y"))
}
