% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.R
\name{extract_params}
\alias{extract_params}
\title{extract and assign default function arguments to the global environment}
\usage{
extract_params(func, verbose = FALSE)
}
\arguments{
\item{func}{A function}

\item{verbose}{A logical value. If TRUE, the function will print the default
parameter values to the console.}
}
\value{
Invisibly returns NULL. Assigns the default values of the function
to the global environment.
}
\description{
This function extracts the default arguments of a given function and assigns
them to variables in the global environment. It is particularly useful for
debugging when you need quick access to the default values.
}
\note{
This function modifies the global environment by assigning variables.
This is intentional for debugging purposes but should be used with caution
in production code.
}
\examples{
rejection_sampler <- function(D, n_trails = 10000, seed = 2024) {}
extract_params(rejection_sampler)
print(n_trails)
# 10000
}
