% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-rnaseq.R
\name{filter_calcpm_dge}
\alias{filter_calcpm_dge}
\title{filter countData by gene expression, perform cpm and log2(cpm+1) transformation}
\usage{
filter_calcpm_dge(count_df, group = NULL, min_count = 10, min_prop = 0.1)
}
\arguments{
\item{count_df}{a data.frame of count data, rows are genes and columns are samples}

\item{group}{a vector of group information, default is NULL, every sample is in the same group}

\item{min_count}{minimum count for a gene to be kept, default is 10}

\item{min_prop}{minimum proportion of samples in the smallest group to have the minimum count, default is 0.1}
}
\value{
a list of edgeR objects, including \code{dge_count}, \code{dge_cpm}, \code{dge_keep}, \code{dge_keep_cpm}, \code{dge_keep_cpm_log}
}
\description{
filter countData by gene expression, perform cpm and log2(cpm+1) transformation
}
\examples{
filter_calcpm_dge(count_df)
}
