% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{find_project_root}
\alias{find_project_root}
\title{Find a project root by walking upward}
\usage{
find_project_root(anchor = ".", marker = "DESCRIPTION")
}
\arguments{
\item{anchor}{A character string specifying the path to start from. If
`anchor` is a file path, the search starts from its parent directory.}

\item{marker}{A character string specifying the project marker to search for.}
}
\value{
The absolute path to the directory containing \code{marker}.
}
\description{
Walk upward from an anchor path until a project marker file or directory is
found.
}
\examples{
\dontrun{
find_project_root(marker = "DESCRIPTION")
}
}
