% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-rnaseq.R
\name{get_top_var_mat}
\alias{get_top_var_mat}
\title{get the the var_gene_list and topN (or top percent) most variable genes}
\usage{
get_top_var_mat(count_df, prop = 0.2, topN = NULL)
}
\arguments{
\item{count_df}{a data.frame of count data, rows are genes and columns are samples}

\item{prop}{proportion of top genes to return, default is 0.2}

\item{topN}{number of top genes to return, default is NULL}
}
\value{
a list of variance (\code{var_genes_list}) and the expression matrix of most-variable genes (\code{topN_mat})
}
\description{
get the the var_gene_list and topN (or top percent) most variable genes
}
\examples{
get_top_var_mat(count_df)
}
