% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngs-rnaseq.R
\name{log_transform}
\alias{log_transform}
\title{perform log transformation with psuedo count}
\usage{
log_transform(df_mat, count = 1, base = 2)
}
\arguments{
\item{df_mat}{data.frame or matrix}

\item{count}{psuedo count, default is 1}

\item{base}{base of the log transformation, default is 2}
}
\value{
a log transformed data.frame
}
\description{
perform log transformation with psuedo count
}
\examples{

log_transform(count_df)
}
