% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_dir.R
\name{make_dir}
\alias{make_dir}
\title{Title: make_dir}
\usage{
make_dir(dir)
}
\arguments{
\item{dir}{a character string of directory path}
}
\value{
a directory created if it does not exist
}
\description{
Title: make_dir
}
\examples{
make_dir("~/test/")
}
