% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{mkdir}
\alias{mkdir}
\title{\code{mkdir}}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{a character string of directory path}
}
\value{
a directory created if it does not exist
}
\description{
create a directory if it does not exist
}
\examples{
mkdir("~/test/")
}
