% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/game_2048.R
\name{play_2048}
\alias{play_2048}
\title{Play the 2048 game in the R console}
\usage{
play_2048(board_size = 4, target = 2048)
}
\arguments{
\item{board_size}{Board size (default 4)}

\item{target}{Winning tile value (default 2048)}
}
\value{
Final board (invisibly)
}
\description{
Use WASD keys to move tiles.
}
