% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print_pval}
\alias{print_pval}
\title{\code{print_pval}}
\usage{
print_pval(pvalues, threshold = 1e-04, rm_tailing0 = TRUE)
}
\arguments{
\item{pvalues}{a numeric vector of p-values}

\item{threshold}{threshold below which p-values are printed in scientific notation, default is 1e-4}

\item{rm_tailing0}{remove tailing 0 in the decimal, default is TRUE}
}
\value{
a string of p-values
}
\description{
print p-values in a more readable format
}
\examples{
print_pval(0.0101)
}
