% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{vec2print}
\alias{vec2print}
\title{\code{vec2print}}
\usage{
vec2print(vec, lang = "R")
}
\arguments{
\item{vec}{the vector to be printed}

\item{lang}{the language ('R', 'shell', 'python') used, default is 'R'}
}
\value{
a string, which can be copy and paste to the code chunk
}
\description{
print a vector so that the output can be copy/paste to the code chunk
}
\examples{
vec2print(c("Apple", "Orange"))
}
