% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_df}
\alias{count_df}
\title{the example count matrix in data.frame format}
\format{
An object of class \code{data.frame} with 100 rows and 20 columns.
}
\usage{
count_df
}
\description{
the example count matrix in data.frame format
}
\keyword{datasets}
