% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_df}
\alias{count_df}
\title{Example count matrix}
\format{
A data.frame with rows as genes and columns as samples.
}
\usage{
count_df
}
\description{
A data.frame containing example count data.
}
\keyword{datasets}
