% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meth_df}
\alias{meth_df}
\title{the example methylation level (beta value) matrix in data.frame format}
\format{
An object of class \code{data.frame} with 100 rows and 20 columns.
}
\usage{
meth_df
}
\description{
the example methylation level (beta value) matrix in data.frame format
}
\keyword{datasets}
