% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{meth_df}
\alias{meth_df}
\title{Example methylation level (beta value) matrix}
\format{
A data.frame with rows as features and columns as samples.
}
\usage{
meth_df
}
\description{
A data.frame containing example methylation beta values.
}
\keyword{datasets}
