% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory.R
\name{mkdir}
\alias{mkdir}
\title{create a directory if it does not exist}
\usage{
mkdir(dir)
}
\arguments{
\item{dir}{A character string specifying the path of the directory to be checked and created if necessary.}
}
\value{
The specified directory path, with the directory created if it did not exist.
}
\description{
This function checks whether a directory exists at the specified path. If the directory does not exist, it is created.
}
\examples{
mkdir("~/test/")
}
