% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{save2pdf}
\alias{save2pdf}
\title{\code{save2pdf}}
\usage{
save2pdf(file = NA, plot_code = NA, overwrite = FALSE, append = TRUE, ...)
}
\arguments{
\item{file}{a character string of the file path}

\item{plot_code}{a language object of the plot code, usually use \code{quote} to wrap the code}

\item{overwrite}{a logical value, whether to overwrite the existing file, default is FALSE}

\item{append}{a logical value, whether to append the new plot to the existing file, default is TRUE}

\item{...}{other parameters passed to \code{pdf}}
}
\value{
a pdf file
}
\description{
save a plot to a pdf file
}
\examples{
save2pdf(file = "mytest_plot.pdf", width = 4, height = 4, overwrite = FALSE,
                   plot_code = quote({
                     t = seq(0, 100, 1)
                     plot(cos(t) + t*sin(t), sin(t) - t* cos(t), type = "l")
                   }))
}
